Fixed issues with autonomous
- Fixed all instances of .auto() to be .trajectory which takes in a SwerveAutoBuilder
- Modified .buildAuto() and .fromStringToCommand() to construct and use a SwerveAutoBuilder.
- Added builder.resetPose() to first trajectory command returned by fromStringToCommand()
- Made sure to set navx when starting true auto
- Cry a wee bit

+++ src/main/java/org/team555/components/managers/MiscData.java
@@ -26,6 +26,6 @@ public class MiscData extends ManagerBase
             REVPhysicsSim.getInstance().run();
         }
 
-        lastFps = fpsFilter.calculate(1 / CommandRobot.deltaTime());
+        lastFps = fpsFilter.calculate((int)(1 / CommandRobot.deltaTime()));
     }
 }
